// build-test/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface DeleteEnvirmentList {
    id: number;
  }

  interface CreateAndUpdateEnvList {
    code: string;
    id: number;
    name: string;
    remark: string;
    is_sys?: boolean;
    create_time?: string;
    update_time?: string;
    sync_env_id?: number | string;
  }
// ...
}